# frontend/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
tags:
- name: Language
  description: Api to manage languages
paths:
  /api/enrollments:
    post:
      tags:
      - enrollment-controller
      summary: Enroll user into a language
      description: Creates a UserLanguage record linking a user and a language.
      operationId: createEnrollment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentRequest"
        required: true
      responses:
        "204":
          description: Enrollment created successfully
        "404":
          description: User or Language not found
        "409":
          description: User already enrolled in language
  /api/languages:
    get:
      tags:
      - Language
      summary: List of all languages
      description: It returns a list of all the available languages
      operationId: getLanguages
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LanguageResponse"
  /api/cefrs:
    get:
      tags:
      - cefr-controller
      operationId: getAllCefrs
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CefrResponse"
components:
  schemas:
    CreateEnrollmentRequest:
      type: object
      properties:
        languageId:
          type: integer
          format: int64
        starterMinutes:
          type: integer
          format: int32
          minimum: 1
        dailyGoalMinutes:
          type: integer
          format: int32
          minimum: 1
      required:
      - dailyGoalMinutes
      - languageId
      - starterMinutes
    LanguageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Internal id
          example: 1
        name:
          type: string
          description: Display name
          example: Spanish
        code:
          type: string
          description: Language code
          example: es
          maxLength: 10
      required:
      - code
      - id
      - name
    CefrResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        level:
          type: string
        minutes:
          type: integer
          format: int32
        hours:
          type: integer
          format: int32
        words:
          type: integer
          format: int32
        descriptionKey:
          type: string
        order:
          type: integer
          format: int32